Moved jobguard to the development directory.
